Rename and Refactor: Modularized the code that creates the dataset and moved it under the Module folder
