+++ module/gendata/save_data.py
@@ -1,11 +1,9 @@
 import pandas as pd
 
-def save_data(data, filename):
+def save_jsonl(data, filepath):
     df = pd.DataFrame(data)
-    df.to_csv(f'module/gendata/files/{filename}.csv', encoding='utf-8-sig')
-    df.to_json(f'module/gendata/files/{filename}.json', force_ascii=False, orient='index', indent=4)
-    df.to_json(f"module/gendata/files/{filename}.jsonl", orient='records', lines=True, force_ascii=False)
+    df.to_json(f"{filepath}.jsonl", orient='records', lines=True, force_ascii=False)
 
-def save_log_data(data, filename):
+def save_csv(data, filepath):
     df = pd.DataFrame(data)
-    df.to_csv(f'module/gendata/files/{filename}.csv', encoding='utf-8-sig')
+    df.to_csv(f'{filepath}.csv', encoding='utf-8-sig')
